fix: lint warnings
