+++ src/api/common.ts
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
 import {
   QueryClient,
   UseQueryOptions,
