+++ src/components/Discounts/DiscountDetailRow.tsx
@@ -1,12 +1,12 @@
 import { format } from "date-fns";
-import { Badge, Button, Icon } from "design-react-kit";
+import { Button, Icon } from "design-react-kit";
 import { Fragment, useMemo, useState } from "react";
 import { useHistory } from "react-router-dom";
 import { Row } from "react-table";
 import {
+  Agreement,
   BucketCodeLoadStatus,
   Discount,
-  DiscountState,
   Profile
 } from "../../api/generated";
 import EditIcon from "../../assets/icons/edit.svg?react";
@@ -21,121 +21,17 @@ import MultilanguageProfileItem from "../Profile/MultilanguageProfileItem";
 import ProfileItem from "../Profile/ProfileItem";
 import { getEditDiscountRoute } from "../../navigation/utils";
 import ImportationStatus from "./ImportationStatus";
+import { getDiscountComponent } from "./getDiscountComponent";
 
 type Props = {
   row: Row<Discount>;
-  agreement: any;
+  agreement: Agreement;
   onPublish: () => void;
   onUnpublish: () => void;
   onDelete: () => void;
   onTest: () => void;
   profile?: Profile;
   maxPublishedDiscountsReached: boolean;
-};
-
-export const getDiscountComponent = (state: DiscountState) => {
-  switch (state) {
-    case "draft":
-      return (
-        <Badge
-          className="fw-normal"
-          pill
-          tag="span"
-          style={{
-            backgroundColor: "white",
-            color: "#5C6F82",
-            border: "1px solid #5C6F82"
-          }}
-        >
-          Bozza
-        </Badge>
-      );
-
-    case "published":
-      return (
-        <Badge className="fw-normal" color="primary" pill tag="span">
-          Pubblicata
-        </Badge>
-      );
-
-    case "suspended":
-      return (
-        <Badge
-          className="fw-normal"
-          pill
-          tag="span"
-          style={{
-            backgroundColor: "#EA7614",
-            border: "1px solid #EA7614",
-            color: "white"
-          }}
-        >
-          Sospesa
-        </Badge>
-      );
-
-    case "expired":
-      return (
-        <Badge
-          className="fw-normal"
-          pill
-          tag="span"
-          style={{
-            backgroundColor: "white",
-            border: "1px solid #C02927",
-            color: "#C02927"
-          }}
-        >
-          Scaduta
-        </Badge>
-      );
-
-    case "test_pending":
-      return (
-        <Badge
-          className="fw-normal"
-          pill
-          tag="span"
-          style={{
-            backgroundColor: "white",
-            border: "1px solid #EA7614",
-            color: "#EA7614"
-          }}
-        >
-          In test
-        </Badge>
-      );
-    case "test_failed":
-      return (
-        <Badge
-          className="fw-normal"
-          pill
-          tag="span"
-          style={{
-            backgroundColor: "white",
-            border: "1px solid #C02927",
-            color: "#C02927"
-          }}
-        >
-          Test fallito
-        </Badge>
-      );
-    case "test_passed":
-      return (
-        <Badge
-          className="fw-normal"
-          pill
-          tag="span"
-          style={{
-            backgroundColor: "white",
-            border: "1px solid #008255",
-            color: "#008255"
-          }}
-        >
-          Test superato
-        </Badge>
-      );
-  }
 };
 
 const DiscountDetailRow = ({
