+++ src/components/Discounts/TestModal.tsx
@@ -2,9 +2,9 @@ import { Button } from "design-react-kit";
 import { Modal, ModalHeader, ModalBody, ModalFooter } from "reactstrap";
 
 type Props = {
-  isOpen: any;
-  toggle: any;
-  testRequest: any;
+  isOpen: boolean;
+  toggle(): void;
+  testRequest(): void;
 };
 
 const TestModal = ({ isOpen, toggle, testRequest }: Props) => (
