+++ src/components/Discounts/getDiscountComponent.tsx
@@ -1,0 +1,107 @@
+import { Badge } from "design-react-kit";
+import { DiscountState } from "../../api/generated";
+
+export const getDiscountComponent = (state: DiscountState) => {
+  switch (state) {
+    case "draft":
+      return (
+        <Badge
+          className="fw-normal"
+          pill
+          tag="span"
+          style={{
+            backgroundColor: "white",
+            color: "#5C6F82",
+            border: "1px solid #5C6F82"
+          }}
+        >
+          Bozza
+        </Badge>
+      );
+
+    case "published":
+      return (
+        <Badge className="fw-normal" color="primary" pill tag="span">
+          Pubblicata
+        </Badge>
+      );
+
+    case "suspended":
+      return (
+        <Badge
+          className="fw-normal"
+          pill
+          tag="span"
+          style={{
+            backgroundColor: "#EA7614",
+            border: "1px solid #EA7614",
+            color: "white"
+          }}
+        >
+          Sospesa
+        </Badge>
+      );
+
+    case "expired":
+      return (
+        <Badge
+          className="fw-normal"
+          pill
+          tag="span"
+          style={{
+            backgroundColor: "white",
+            border: "1px solid #C02927",
+            color: "#C02927"
+          }}
+        >
+          Scaduta
+        </Badge>
+      );
+
+    case "test_pending":
+      return (
+        <Badge
+          className="fw-normal"
+          pill
+          tag="span"
+          style={{
+            backgroundColor: "white",
+            border: "1px solid #EA7614",
+            color: "#EA7614"
+          }}
+        >
+          In test
+        </Badge>
+      );
+    case "test_failed":
+      return (
+        <Badge
+          className="fw-normal"
+          pill
+          tag="span"
+          style={{
+            backgroundColor: "white",
+            border: "1px solid #C02927",
+            color: "#C02927"
+          }}
+        >
+          Test fallito
+        </Badge>
+      );
+    case "test_passed":
+      return (
+        <Badge
+          className="fw-normal"
+          pill
+          tag="span"
+          style={{
+            backgroundColor: "white",
+            border: "1px solid #008255",
+            color: "#008255"
+          }}
+        >
+          Test superato
+        </Badge>
+      );
+  }
+};
