+++ src/components/Form/FormContainer.tsx
@@ -1,8 +1,9 @@
+import { ReactNode } from "react";
 import Container from "../Container/Container";
 
 type Props = {
   className?: string;
-  children: any;
+  children: ReactNode;
 };
 
 const FormContainer = ({ className = "", children }: Props) => (
