+++ src/components/Header/LogoutModal.tsx
@@ -2,9 +2,9 @@ import { Button } from "design-react-kit";
 import { Modal, ModalHeader, ModalBody, ModalFooter } from "reactstrap";
 
 type Props = {
-  isOpen: any;
-  toggle: any;
-  logout: () => void;
+  isOpen: boolean;
+  toggle(): void;
+  logout(): void;
 };
 
 const LogoutModal = ({ isOpen, toggle, logout }: Props) => (
