+++ src/components/Layout/Layout.tsx
@@ -1,9 +1,10 @@
+import { ReactNode } from "react";
 import Header from "../Header/Header";
 import Footer from "../Footer/Footer";
 
 type Props = {
   hasHeaderBorder?: boolean;
-  children: any;
+  children: ReactNode;
 };
 
 const Layout = ({ hasHeaderBorder = false, children }: Props) => (
