+++ src/types/react-table-config.d.ts
@@ -64,6 +64,7 @@ declare module "react-table" {
       // note that having Record here allows you to add anything to the options, this matches the spirit of the
       // underlying js library, but might be cleaner if it's replaced by a more specific type that matches your
       // feature set, this is a safe default.
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
       Record<string, any> {}
 
   export interface Hooks<
@@ -115,7 +116,7 @@ declare module "react-table" {
 
   export interface Cell<
     D extends Record<string, unknown> = Record<string, unknown>,
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars, @typescript-eslint/no-explicit-any
     V = any
   > extends UseGroupByCellProps<D>,
       UseRowStateCellProps<D> {}
